requirements.txt
Pillow>=9.0.0
exifread>=3.0.0
moviepy>=1.0.3
eyed3>=0.9.7
python-magic>=0.4.27
rich>=13.0.0
ffmpeg-python>=0.2.0
pdfplumber>=0.9.0
python-docx>=0.8.11
mutagen>=1.47.0
flask>=3.0.0 